requirements.txt
pyaudio
librosa
matplotlib
numpy
soundfile
dtw-python
scipy
simpleaudio
streamlit
streamlit-audio-recorder
streamlit-webrtc
pandas
requests
yt-dlp
openai-whisper
torch
torchaudio
transformers
arabic-reshaper
pyarabic
python-bidi
gTTS 